fix(vc): address critical+high issues from pass-4 review
1. CRITICAL: Add missing `remote_signer` field to ValidatorClient struct
   and free it in deinit(). Without this field the assignment at line ~490
   was writing to a nonexistent struct member.

2. HIGH: Fix bitlist data loss for small committees in parseAggregateAttestation.
   When byte_len == 1 no data bytes were copied. Now we copy the single byte
   and mask out the SSZ sentinel bit so the bitvector contains only the real
   committee bits.

3. HIGH: Move slashing-protection check BEFORE the remote-signer branch in
   signBlock/signAttestation. Remote validators were bypassing the slashing DB
   entirely, opening a slashing window for web3signer-backed keys.

4. HIGH: Change block_delay_sec from u64 to u32 in import_block.zig to match
   the expected parameter type of onBlockFromState.

+++ src/chain/blocks/import_block.zig
@@ -194,7 +194,7 @@ pub fn importVerifiedBlock(
             // Sync/API/regen blocks use delay=5 (> ATTESTATION_DUE threshold of 4s)
             // so they do NOT receive proposer boost.
             // TODO: replace with actual wall-clock arrival time tracking.
-            const block_delay_sec: u64 = switch (block_input.source) {
+            const block_delay_sec: u32 = switch (block_input.source) {
                 .gossip => 0, // timely — proposer boost applies
                 else => 5, // late — no proposer boost (threshold is >4s)
             };
